improve nav structure, css smarts for docs
sitemap and gemini link updates, readme touchups

Coding agents

dialog refactor

remove Sync Rate Limit nudge
