+++ docs/_includes/sidebar.html
@@ -17,11 +17,11 @@
       {% for item in section.children %}
       {% if item.children %}
       {% for grandchild in item.children %}
-      {% if page.url contains grandchild.url %}
+      {% if page.url == grandchild.url %}
       {% assign section_active = true %}
       {% endif %}
       {% endfor %}
-      {% elsif page.url contains item.url %}
+      {% elsif page.url == item.url %}
       {% assign section_active = true %}
       {% endif %}
       {% endfor %}
@@ -33,7 +33,7 @@
             {% if item.children %}
             {% assign item_active = false %}
             {% for grandchild in item.children %}
-            {% if page.url contains grandchild.url %}
+            {% if page.url == grandchild.url %}
             {% assign item_active = true %}
             {% endif %}
             {% endfor %}
@@ -42,14 +42,14 @@
               <ul class="nav-grandchildren">
                 {% for grandchild in item.children %}
                 <li>
-                  <a href="{{ grandchild.url | relative_url }}" {% if page.url contains grandchild.url %}class="active"
+                  <a href="{{ grandchild.url | relative_url }}" {% if page.url == grandchild.url %}class="active"
                     {% endif %}>{% if grandchild.icon %}<span class="nav-icon"><img class="icon-light" src="{{ '/assets/icons/' | append: grandchild.icon | append: '-light.svg' | relative_url }}" alt="" width="16" height="16"><img class="icon-dark" src="{{ '/assets/icons/' | append: grandchild.icon | append: '-dark.svg' | relative_url }}" alt="" width="16" height="16"></span>{% endif %}{{ grandchild.title }}</a>
                 </li>
                 {% endfor %}
               </ul>
             </details>
             {% else %}
-            <a href="{{ item.url | relative_url }}" {% if page.url contains item.url %}class="active" {% endif %}>{{
+            <a href="{{ item.url | relative_url }}" {% if page.url == item.url %}class="active" {% endif %}>{{
               item.title }}</a>
             {% endif %}
           </li>
