+++ docs/getting-started/agents/gemini-cli.md
